Use FodUtils for FOD config checking
base: Add our CorvusUtils helper class

From time to time we need to filter out certain features because not every
device is created equal. This makes it easy to call from one place instead
of adding the same code time and time again.

As of right now this class includes checks for

- Fingerprint support
- Fingerprint support & enrollment

Change-Id: Ia27dc8d49a78c843f2a185f3aef195f17cc090aa